chore: rename shared to common
