+++ packages/sshnoports/lib/sshnp/cleanup.dart
@@ -1,7 +1,7 @@
 // dart packages
 import 'dart:io';
 // local packages
-import 'package:sshnoports/shared/utils.dart';
+import 'package:sshnoports/common/utils.dart';
 // atPlatform packages
 import 'package:at_utils/at_logger.dart';
